# api-docs.json
{"openapi":"3.0.1","info":{"title":"Fitness Tracker API","version":"1.0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/v1/users/{userId}":{"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUserById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/v1/trainings/{trainingId}":{"put":{"tags":["training-controller"],"operationId":"updateTraining","parameters":[{"name":"trainingId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertTrainingDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Training"}}}}}}},"/v1/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}},"post":{"tags":["user-controller"],"operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/v1/trainings":{"get":{"tags":["training-controller"],"operationId":"getAllTrainings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainingDto"}}}}}}},"post":{"tags":["training-controller"],"operationId":"addTraining","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertTrainingDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Training"}}}}}}},"/v1/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/v1/users/simple":{"get":{"tags":["user-controller"],"operationId":"getAllUsersSimple","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserSimpleDto"}}}}}}}},"/v1/users/older/{time}":{"get":{"tags":["user-controller"],"operationId":"getUserOlderThan","parameters":[{"name":"time","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"/v1/users/email":{"get":{"tags":["user-controller"],"operationId":"getUserByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"/v1/trainings/{userId}":{"get":{"tags":["training-controller"],"operationId":"getTrainingByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainingDto"}}}}}}}},"/v1/trainings/finished/{afterTime}":{"get":{"tags":["training-controller"],"operationId":"getFinishedTraining","parameters":[{"name":"afterTime","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainingDto"}}}}}}}},"/v1/trainings/activityType":{"get":{"tags":["training-controller"],"operationId":"getTrainingByActivityType","parameters":[{"name":"activityType","in":"query","required":true,"schema":{"type":"string","enum":["RUNNING","CYCLING","WALKING","SWIMMING","TENNIS"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainingDto"}}}}}}}}},"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"birthdate":{"type":"string","format":"date"},"email":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"birthdate":{"type":"string","format":"date"},"email":{"type":"string"}}},"UpsertTrainingDto":{"type":"object","properties":{"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"activityType":{"type":"string","enum":["RUNNING","CYCLING","WALKING","SWIMMING","TENNIS"]},"distance":{"type":"number","format":"double"},"averageSpeed":{"type":"number","format":"double"},"userId":{"type":"integer","format":"int64"}}},"Training":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"activityType":{"type":"string","enum":["RUNNING","CYCLING","WALKING","SWIMMING","TENNIS"]},"distance":{"type":"number","format":"double"},"averageSpeed":{"type":"number","format":"double"}}},"UserSimpleDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"}}},"TrainingDto":{"type":"object","properties":{"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"activityType":{"type":"string","enum":["RUNNING","CYCLING","WALKING","SWIMMING","TENNIS"]},"distance":{"type":"number","format":"double"},"averageSpeed":{"type":"number","format":"double"},"user":{"$ref":"#/components/schemas/User"}}}}}}